+++ docs/changelog.md
@@ -15,8 +15,18 @@
 
 # Project
 
+## workspace
+### examples
+- new example scripts: `scope_guard`, `current_guard`.
+
+## features & flags
+- update the intended scope of the `_docs_examples` feature.
+
 ---
 
 # Modules
+
+## yard
+- update `_use_or_shim!` to add `_doc!` macro support.
 
 [0.28.0]: https://github.com/andamira/devela/releases/tag/v0.28.0